Djikstra's Algorithm Solver
Finds the longest single path in a spanning tree for a graph described by the bottom half of an adjacency matrix. Number of vertices must be within range [1,100]. 
Input A(i,j) is the distance between i and j. If no path exists, 'x' should be given input. A(i,i) should not be provided.
Example Input: 
5
50
30 5
100 20 50
10 x x 10
Output for given input:
35